test: colocate strict tool regressions
