+++ packages/openai-base/tests/tool-converter-strict-fallback.test.ts
@@ -55,37 +55,6 @@ const freeFormMapTool: Tool = {
   },
 }
 
-const anyOfWithOptionalVariantTool: Tool = {
-  name: 'store_variant',
-  description: 'Store a union variant',
-  inputSchema: {
-    type: 'object',
-    properties: {
-      value: {
-        anyOf: [
-          {
-            type: 'object',
-            properties: {
-              kind: { const: 'optional' },
-              note: { type: 'string' },
-            },
-            required: ['kind'],
-          },
-          {
-            type: 'object',
-            properties: {
-              kind: { const: 'nullable' },
-              note: { type: ['string', 'null'] },
-            },
-            required: ['kind', 'note'],
-          },
-        ],
-      },
-    },
-    required: ['value'],
-  },
-}
-
 describe('responses tool converter — strict fallback', () => {
   it('uses strict:true for strict-subset schemas', () => {
     const out = convertFunctionToolToResponsesFormat(strictSafeTool)
@@ -114,15 +83,6 @@ describe('responses tool converter — strict fallback', () => {
     ).toEqual({ type: 'string' })
   })
 
-  it('falls back to strict:false when an anyOf variant needs null widening', () => {
-    const out = convertFunctionToolToResponsesFormat(
-      anyOfWithOptionalVariantTool,
-    )
-    expect(out.strict).toBe(false)
-    expect(
-      (out.parameters as any).properties.value.anyOf[0].required,
-    ).toEqual(['kind'])
-  })
 })
 
 describe('chat-completions tool converter — strict fallback', () => {
@@ -144,12 +104,6 @@ describe('chat-completions tool converter — strict fallback', () => {
     expect(out.function.strict).toBe(false)
   })
 
-  it('falls back to strict:false when an anyOf variant needs null widening', () => {
-    const out = convertFunctionToolToChatCompletionsFormat(
-      anyOfWithOptionalVariantTool,
-    )
-    expect(out.function.strict).toBe(false)
-  })
 })
 
 // This is the converter the provider tool-dispatcher (`convertToolsToProviderFormat`)
@@ -180,10 +134,4 @@ describe('function-tool adapter converter — strict fallback', () => {
     expect(out.strict).toBe(false)
   })
 
-  it('falls back to strict:false when an anyOf variant needs null widening', () => {
-    const out = convertFunctionToolToAdapterFormat(
-      anyOfWithOptionalVariantTool,
-    )
-    expect(out.strict).toBe(false)
-  })
 })
